refactor(devices): extract shared device columns definition
Create shared columns.tsx in DeviceSelectTable component that both
the device list page and device select table can use:

- src/components/DeviceSelectTable/columns.tsx: Shared getDeviceColumns()
- src/pages/devices/list/index.tsx: Uses shared columns + action column
- src/components/DeviceSelectTable/index.tsx: Uses shared columns

This ensures both pages stay in sync when columns are updated.
The device list page adds its own action column for edit/enable/disable/delete.
